Add a class for info-text and start to style
I started to style the info section very basically. This is still
subject to great changes. But, it was noticed that a class was missing
for the info-text so this was added. Also for quality of life the
width and height of each image used previously was specified in the
style sheet not just the html page.