Style the info section
The info section was completely styled such that the images and text
are all wrapped in a flexbox in order to have such a style. These items
were then wrapped in their own individual flexboxes in order to maintain
the text centered directly under the images. Also, the elements can wrap
and the subtext color was changed slightly to make the text more
readable.