Style the sign up section
I styled the sign up section using a flexbox for the main blue box. Then
I used another flexbox around the text by adding a new div with the
class name of sign-up-box as it wraps the text. Then new styling was
added to this sign up button such that it stands out from the box, and
does not blend in with it.