Style the footer and add Roboto font import
I styled the footer by making the footer div a flexbox to properly
display the text. Then I styled the footer text in order to get the
correct style. Afterwards when the project was completed I added an
import to change the font family to that of google's roboto font using
a universal tag in css to do this, and by importing it in the
index.html file.